Renamed test modules to match their package name
